+++ src/Instrument.cpp
@@ -1,21 +1,21 @@
 #include <Instrument.h>
 
-int Instrument::getSettingsCount() const
+uint8_t Instrument::getSettingsCount() const
 {
     return 0;
 }
 
-void Instrument::incrSetting(int i)
+void Instrument::incrSetting(uint8_t i)
 {
     this->configureSetting(i, this->getSettingValue(i) + 1);
 }
 
-void Instrument::decrSetting(int i)
+void Instrument::decrSetting(uint8_t i)
 {
     this->configureSetting(i, this->getSettingValue(i) - 1);
 }
 
-String Instrument::logSetting(int i)
+String Instrument::logSetting(uint8_t i)
 {
     return String(getSettingName(i)) + ":" + getSettingValue(i);
 }
